language icon by Onur Mustak Cobanli
Taken from http://www.languageicon.org, there distributed under CC
license. Quote from the website, 2021-10-07:

Can I use the Language Icon Freely?

Yes, you can use the Language Icon for commercial and non-commercial
projects; A' Design Award & Competition, who has organized the
Language Icon design competition, Farhat Datta, the creator of the new
Language Icon, and Onur Mustak Cobanli, the creator of the original
idea hereby declare that the Language Icon is a collaborated creation
released with a CC license with following terms:
Relax-Attribution. You are suggested but not required to attribute the
work when using for internet / digital use. You must attribute it in
any other use which is not on internet. Attribute to: A’ Design Award
& Competition, Onur Müştak Çobanlı and Farhat Datta with URL
http://www.languageicon.org for non-internet usage.  Share Alike. If
you alter, transform, or build upon this work, you may distribute the
resulting work only under the same or similar license to this one,
attribute to original author, resulting work cannot be
commercial. Semi-Noncommercial. You may not use this work as a central
element (or one of the core elements) for commercial purposes; Such as
a design on a tshirt, on a book cover etc unless you attribute to the
author. On the other hand you can use freely in your websites without
attribution to signify language. Color-Firendly: You can change the
colors as you like, keep the form intact please; scale
proportionally. You can reverse the color or grayscale it if required.